Started working on items logic.

+++ src/main/java/com/ely/spring_simple_rpg_system/entity/Item.java
@@ -27,5 +27,5 @@ public class Item {
     @Column(nullable = false)
     private ItemUsages itemUsage;
 
-    private int effectValue;
+    private Long effectValue;
 }
